refactor(example): split turnkey community page
