+++ examples/turnkey-community-example/src/hooks/turnkey-community-example/use-eth-balance.ts
@@ -1,0 +1,11 @@
+import { useQuery } from "@tanstack/react-query";
+import type { PublicClient } from "viem";
+import type { Address } from "viem";
+
+export function useEthBalance(publicClient: PublicClient | null, walletAddress: Address) {
+  return useQuery({
+    queryKey: ["ethBalance", walletAddress],
+    enabled: !!publicClient,
+    queryFn: async () => publicClient!.getBalance({ address: walletAddress }),
+  });
+}
